Add hidden prefixes
JSTs ignored prefixes are still pushed into the output file sink, which
is not needed when the input and output sources are the same.
The commit adds a stronger version of ignored prefixes, "hidden"
prefixes, which are not processed or pushed to the output sink.

This mainly aims to prevent potential write permission issues on
read-only files during in-place transformation.