Generate stubs for generic arguments (#53)
Until now, generic arguments wouldn't have stubs generated, meaning that you could not inject an interface with a type parameter that references your own classes. This fixes that.
For this to be fixed, an exception from the "verbatim copy" rule had to be added: inner classes must use $. This change does not however break any existing *working* builds.